move app and single quote
